Refactor apimodel Structs to Better model structs.
Instead of having apimodel and model duplicated and a lot of conversion
between the two, this kind of merges them back into one overhauled
version. The changes touch everything so it's likely to break things.

+++ app/model/datapoint.go
@@ -4,7 +4,7 @@ package model
 type DataPoint struct {
 	LocalTime string  `json:"label"`
 	EpochTime int64   `json:"x"`
-	Y         int     `json:"y"`
+	Y         float32 `json:"y"`
 	Value     float32 `json:"value"`
 	Tag       string  `json:"tag"`
 }
